Extract RootDrawable and WrappingUtils from GenericHierarchy
